oncology-vector-backend/requirements.txt
fastapi==0.95.2
pydantic==1.10.8
langchain==0.0.156
pymongo[srv]==4.5.0
bson==0.5.10
requests==2.31.0
faiss-cpu  # For CPU version of FAISS
uvicorn==0.22.0         # For running the FastAPI server
pymupdf 
pdfplumber
TTS